Systeme de zombie + changements stats
le zombie est fini à peu pres, faut que je modifie 2-3 trucs et la ca part sur le systeme global de spawn

+++ Undead Rift/Assets/Scripts/Gun.cs
@@ -49,7 +49,7 @@ public class Gun : MonoBehaviour
                 {
                     IDamageable damageable = hitInfo.transform.GetComponent<IDamageable>();
                     if (damageable != null) { 
-                        damageable.TakeDamage(gunData.damage);
+                        damageable.TakeDamage((int)gunData.damage);
                     }
                     GameObject ImpactGO = Instantiate(ImpactEffect,hitInfo.point, Quaternion.LookRotation(hitInfo.normal));
                     Destroy(ImpactGO, 2f);
